Dependencies for analysis notebooks, caching of the NewsFeed, and more prompt to manually provide output schema.
I think additional work is needed to make the output schema part more efficient. I had to add it into the prompt directly to support Anthropic. perhaps I should just not support it. But having the schema in two places presumably increases the cost of running this experiment on platforms that support the OpenAI API response schema mechanism.